test: extract dynamic paths suite
