+++ test/fixtures/dynamic-paths/app/assets/global.css
@@ -1,0 +1,3 @@
+:root {
+  --public-asset: url('/css-only-public-asset.svg');
+}
