+++ test/fixtures/dynamic-paths/app/pages/assets.vue
