+++ test/fixtures/dynamic-paths/app/pages/index.vue
@@ -1,0 +1,3 @@
+<template>
+  <div>index</div>
+</template>
