+++ test/fixtures/dynamic-paths/app/pages/navigate-to.vue
@@ -1,0 +1,7 @@
+<template>
+  <div>You should not see me</div>
+</template>
+
+<script setup>
+await navigateTo('/', { replace: true, redirectCode: 301 })
+</script>
