+++ test/fixtures/dynamic-paths/app/pages/url.vue
@@ -1,0 +1,9 @@
+<script setup>
+const url = useRequestURL()
+</script>
+
+<template>
+  <div>
+    <div>path: {{ url.pathname }}</div>
+  </div>
+</template>
