+++ test/matrix.ts
@@ -40,6 +40,12 @@ export const runsOnceInMatrix = !isMatrixRun || (isCanonicalCombo && isBuilt)
  */
 export const runsOncePerEnvInMatrix = !isMatrixRun || isCanonicalCombo
 
+/**
+ * Like `runsOnceInMatrix` but keeps the builder axis: true in exactly one built project per
+ * builder. Use for builder-sensitive suites that do not depend on other matrix axes.
+ */
+export const runsOncePerBuilderInMatrix = !isMatrixRun || (isBuilt && !asyncContext && isTestingAppManifest)
+
 export const isNuxtPrepare = process.argv.slice(2).includes('prepare')
 
 export function withMatrix (config: NuxtConfig) {
